Don't alarm on "Email in use" response from Identity /guest endpoint
Sometimes POST /guest returns "Email in use" even though we already
checked whether the user existed by looking them up with GET /user
passing their email address. According to identity occasionally users
get into a weird state and this is a result of that. I think this should
still result in a 500 (it's not related to something the user has done)
but we don't want to alarm on it.