Add a higher threshold alarm which we can use for known checkout failure reasons (#6749)
This PR adds a new higher threshold alarm/metric. In #6745 we stopped alarming
on the "email already taken" error from identity. It's a known issue and we
don't want to spend time investigating every time it happens. However, we would
like to know if it starts occurring at a higher volume. This PR adds the
concept of a higher threshold alarm/metric which we can use to cover this case.
The alarm will trigger if 10 or more minutes in the last hour have errors.