Add server side validation for billing address postcode (#6759)
We've had an ongoing issue where requests where the payment method is
PayPal make it to support workers with an email address in the billing
postcode field. We're not sure why this happens but are hoping some
server side validation will prevent it and alert users that they need to
fix something. It primarily seems to affect USD Supporter Plus and
recurring contributions (we collect the postcode for these) so I've
focused on fixing those cases, in the PaidProductValidation object used
by both products.

I've introduced a concept of a preservable error message, i.e. something
we can map to copy to show the user. It'd be nice to extend this but for
now it's a one off for this field. As it stands, it's not perfect - if
there are other validation errors we'll lose the message. I'd like to
fix that in a future PR.