Force the recaptcha to re-initialise after a server side failure
Reinitialise the recaptcha each time server side validation fails. This
causes a new Stripe payment intent to be created.

Before this change, if the reader was paying with Stripe, and server side
validation rejected the call to /subscribe/create, then even if the user
then fixed the issue which caused the validation failure, future form
submissions fail with the following error from Stripe:
[setup_intent_unexpected_state].

[setup_intent_unexpected_state]:
https://stripe.com/docs/error-codes/setup-intent-unexpected-state