Get tsc and lint passing
The versions of react, react-dom and their type definitions seem to need
to be very exact, I've locked them all down in the package.json to
reflect the versions we already had.

Also add immer as we rely on it (and pnpm isn't happy with it being
brought in as a phantom dependency).