Remove unused playwright code and config (#6819)
Removing unused playwright code and config. It's confusing as there's
duplication in the tests which are defined, I think this is a result of having
split the tests in two (cron and smoke) and remnants of the old approach being
left behind.  There are also playwright config files which aren't used or
needed.