Make payment-success metric for Stripe more granular (#6835)
The payment-api emits a payment-success metric from the StripeBackend.
Before, the payment-provider dimension was always hard coded to Stripe.
This change makes the payment-provider dimension more granular so we'll
send one of:

* Stripe
* StripeApplePay
* StripePaymentRequestButton

The reason I want to do this is to have separate alarms for each of
these when we go a period without any successful payments.