Migrate support-frontend from yarn to pnpm (#6806)
We need to move off of yarn v1 as it's deprecated. Using pnpm is generally our
preference for JS/TS projects as it has a number of DevX benefits (better
performance, improvements in monorepos).

The compressed-size-action build check doesn't work on this PR, as the current
branch uses pnpm and the target is still on yarn, which is a scenario it
(understandably) doesn't support.

Using pnpm seems to offer much better behaviour around the
@guardian/support-service-lambdas package compared to yarn 1 which made it hard
to update, requiring us to add a script to do this. It was non-obvious what was
happening when things weren't working for developers, especially those outside
the team.

The package.json has changed in a few places. pnpm is much stricter around
"phantom dependencies" i.e. those which you rely on/import but which aren't
explicitly added to the package.json. Yarn would allow these if they were
available as transitive dependencies, i.e. a dependency of another package
which we do have in our package.json.

I also locked down to specific versions of react and react-dom (and their
definitely typed counterparts). This was because I wanted to keep these
versions identical as we migrated (I ended up getting mismatches which caused
type errors).

I've commented out the caching config for dependencies in the workflow config.
This was causing confusing behaviour as I worked on migrating. Once we're moved
over to pnpm I think we can add it back. Interestingly it doesn't seem to have
made the build any slower on this branch compared to recent builds on main.