Stripe Hosted Checkout switches
I've added a switch under RecurringPaymentMethodSwitches (used by the
client) and SubscriptionsPaymentMethodSwitches (used by the server). In
future we should tidy these up so that client and server use the same
switch. The current setup supports confusing combinations (enabled on
the client but not on the server) which would mean users couldn't check
out.