Re-implement the bigquery-acquisitions-publisher lambda in TypeScript (#6787)
There’s an issue with the compressed jar file size which is close to the limit
of S3 based lambda deployments enforced by AWS. When we tried to update the
google-cloud-bigquery Scala library this pushed us over the size limit. We
could migrate to deploying using ECR images instead of S3, which has a much
higher size limit. But we’d like to migrate to TypeScript anyway, and this will
probably sidestep the size issue due to the smaller size of TypeScript/JS code
artefacts.

Note: we haven't deleted all the old Scala code in this PR as it's been useful
to have it side by side with the new TypeScript. We'll raise a follow up PR to
do this once this PR is merged.

We've tested by going through the checkout process in CODE and inspecting the
resulting row(s) in fact_acquisition_events.