Stripe Hosted Checkout: Validate form data and redirect to Stripe (#6863)
This PR implements the first part of the Stripe hosted checkout flow. For
HomeDelivery/SubscriptonCard Sunday plan selecting the `Credit/Debit card`
payment option and clicking `Continue to payment` will:

* Validate the user form data
* Create the checkout session with Stripe
* Redirect the user to Stripe (client side redirect)
* If the Stripe checkout is successful, redirect back to the checkout with the
  checkout session ID set as a query string arg

This is not a complete flow. In prod the Stripe hosted checkout option will be
hidden by a switch until ready. We also won't be showing any cards directing
the user to the checkout for Sunday.

Missing elements:

* Persisting the user's details on the form before they are redirected to Stripe
* Processing the sub when the user returns to the support site

I'm currently using our default Stripe account. In a subsequent PR I'll migrate
to using a different Stripe account before any of this is available publicly.