Persist and retrieve persisted form fields on checkout (#6891)
For the Stripe hosted checkout flow the checkout process looks like this:

1. Normal checkout form
2. Stripe hosted checkout
3. Back to support site

Validation of form fields etc takes place in the first step, before
transferring to Stripe. When the user returns to our site, we don't want them
to have to fill in their details again, but these details are needed to submit
the create subscription request to support-workers. It's also possible that
something may go wrong with the support-workers execution, in which case we'd
want to place the user on the checkout form, with an error message, but their
form fields still filled.

In this PR we persist the filled form fields to session storage under the key
checkoutSessionFormData, with an expiration of 24 hours (managed by
@guardian/libs). The reason we use 24 hours is to align with the expiration of
the checkout session on Stripe's side. Since we're using session storage this
data is tied to the current browser tab so the expiration will be the shorter
of 24 hours and the lifespan of the current browser tab. After the user
successfully returns from Stripe, we retrieve the details if there's a
checkoutSessionId query string arg. If the checkout session ID doesn't match,
the persisted form details aren't used. If there is a match, and the data is
valid, we fill the form with the persisted details.