Allow persisted deliveryInstructions field to be null
I noticed that when checking out an Observer subscription with
SubscriptionCard that the persisted form details were failing to parse.
This was because deliveryInstructions was set to null (it isn't offered
on the form) rather than undefined. This small schema change means we're
able to de-serialize the persisted json and complete the checkout flow.