Using a wildcard in the product dimension didn't work
It looks like you can't use wildcards with an alarm, only a search
expression. This means we need to enumerate the possible products. Doing
this for both Apple Pay and Google Pay pushes us over the limit for the
number of metrics in the alarm expression (10). So I've split them in
two.