Add alarms for no S+ subscription or recurring contribution with Stripe Express payment methods (#6955)
This PR adds alarms for periods with no Supporter Plus or recurring
contributions paid for by Stripe Express payment methods (i.e. Apple Pay or
Google Pay).

I determined the periods to use for the new alarms by querying the datalake for
the longest period we went without these product/payment method contributions
for the first few months of the year.

The changes include some refactoring to make it easier (and require less code)
to add new alarms for S+ and recurring contributions.