Roll back a major version of @guardian/eslint-config-typescript
We probably want to move off of this to just @guardian/eslint-config in
future, but that requires a jump to eslint 9 which we're not ready for
yet. This reduces the number of eslint violations significantly (and
also sidesteps a performance issue I was seeing).

One warning from eslint currently, I'll see about fixing that
separately.