Re-instate playwright cron test failure alerts (#6974)
The alerting part was commented out a few weeks ago due to flakiness, we
think on the PayPal sandbox end. The tests have continued running and it
looks like they've been much more reliable recently so let's turn the
alerting back on so we find out about any genuine failures.