The tests caught a subtle issue here
I'd changed the behaviour from the previous HOC implementation slightly.
I was passing through all props to the underlying component which
weren't used by Label or Error. But the previous behaviour, which
apparently we relied on, was to pass through _everything_ to the
underlying component, even the props used by Label and Error.