Update snapshots with new emotion generated css class
I couldn't figure out a way to keep these unchanged in this PR. The
approach of setting globalThis.EMOTION_RUNTIME_AUTO_LABEL as we do in
render.ts didn't seem to work. I considered using @emotion/jest as using
it in the snapshot configuration as described [here] but we decided as a
team to avoid introducing an additional dependency and instead to lose
the component name in snapshots.

[here]: https://emotion.sh/docs/testing