Fix support-workers integration tests (#7007)
Fixing a handful of failing integration tests and some infra issues with the
lambda which runs the tests (it was getting the support-workers jar from the
wrong S3 bucket).

The integration tests have been failing for a while but we've not been getting
notified. This is fixed in #7003, but before merging that PR I'd like to get
the tests back to green.