Hosted checkout: verify that user entered form fields match the checkout session (#7002)
When the checkout session is created store the SHA256 hash of the user entered
form fields in the metadata field. When retrieving the checkout session in
support-workers, compare the hash with the hash of the form fields provided. If
they don't match, send a CloudWatch metric event.

There's no scenario where the form fields (which we stash in session storage)
should change between redirecting the user to Stripe and coming back to the
support site to complete the transaction, so let's make sure it doesn't. Before
rolling this out in a way which actually prevents a purchase, let's emit a
metric event so we can see if it ever happens first.