We have to remove any existing checkoutSessionId from the return URL
In the case that you've been to Stripe, come back, and seen an error,
there will have been a checkoutSessionId in the query string. When you
go back to Stripe the new checkout session ID will be appending so we'll
have two! This breaks things so remove before creating the checkout
session.