Handle errors from our backend after returning from the Stripe hosted checkout (#7032)
When a user returns from the Stripe hosted checkout to the generic checkout we
automatically retrieve their form details and submit the form. It's
theoretically possible (though we haven't seen it) that we may return an error
from our backend at this point. If this happened the user would be left on the
form with an error. Resubmitting at this stage will always just resubmit to the
backend and use their Stripe hosted checkout details. Depending on the type of
error this may not be sufficient to fix the problem. Because this is such an
unlikely scenario, we decided the best thing to do is reset the form fields and
effectively start over, which is what this PR adds.

As part of this I've migrated the checkout session retrieval logic to a custom
hook which wraps useState and useEffect. The reason I'm doing this is so that
we can manipulate/clear the checkout session state and cause a re-render as
expected.