Extract sale/transaction ID from PayPal payment and add to acquisition event (#7073)
When a user makes a single contribution with PayPal, extract an additional ID
field from the payment object we get back from PayPal and add it to the
acquisition event under paypalTransactionId. Later, this will be used to
populate a field in Salesforce.

Currently CSRs find it impossible to find a transaction in PayPal if the PayPal
email address doesn't match the identity account email address. This is becuase
the payment ID field on the acquisition event, which is surfaced in Salesforce,
cannot be used to search transactions in PayPal (or rather it can be used, but
doesn't match anything useful).