Gather and report core web vitals (client side router pages only) (#7072)
Start using @guardian/core-web-vitals to gather and report core web vitals
(CWV) data on the support site. Initially I'm only rolling this out to pages
served by the client side router. We can roll out to other pages in future if
this is successful.

This data is made available in the datalake. It's is distinguishable from other
existing CWV data collected by using the platform field which is being set to
support.

This is this so that we can track our CWV performance over time and see the
impact of changes made.

I tested by dialling up the sampling to 100% locally and can see the data
appear in the datalake. All data goes to the production table currently (this
is existing behaviour of @guardian/core-web-vitals and the events collector
pipeline). Non prod data has a stage of CODE.