Use correct GoCardless payment gateway for Sunday paper subs
Since this create payment method lambda was migrated to TypeScript we
lost some functionality around GoCardless payment gateways. For Sunday
paper subs we want to use a different payment gateway.

In the Scala version of this lambda we switched the GoCardless payment
gateway depending on the product, so implement the same logic here.