Make missing Stripe card payment fields on the checkout more obvious to the user (#7099)
When a user selects the Credit/Debit card payment method and submits the form
without entering one or more of the required card fields, show specific errors
next to the field and scroll to them.

Before if the user performed the actions above, it wasn't very clear what had
gone wrong. A very generic error message was show below the pay/submit button
and we didnt't make it clear what the user can do to fix things.

Before this resulted in an exception which is reported to Sentry. The volumes
of this error suggests this situation is actually happening to users.

Now, when this happens we'll show an error message next to the relevant
field(s) and scroll the viewport to the payment section of the form to make
sure these errors are visible.