Show error if recaptcha not set for DirectDebit
Before this change if the user tried to pay with Direct Debit but didn't
check the recaptcha the form submit would fail and we'd render a very
generic error below the submit button, which isn't helpful to the user.

Now when this happens an error is shown by the recaptcha and the
viewport is scrolled to the payment section of the form.

Currently there's an inconsistency in the UI when other Direct Debit
fields are missed. We use native HTML validation. We'd like to move to
the approach we use for Stripe (where the native validation isn't
possible as we don't own the form fields). We hope to address this in a
future PR.