Extract Ad-Lite tests out to start from landing page
This establishes a new pattern where checkout tests start on the landing
page for the relevant product. Eventually this will be applied to all
the current tests defined in checkout.test.ts, which will be grouped by
the landing page.

We'd like to start tests on the landing page as this is closer to how
real users will interact with the site.