Prevent flash of checkout without overlay before redirect to thank you page (#7122)
Move setIsProcessingPayment(false) call to catch block.

Moving this state update to the catch block because it has the effect of
removing the processing overlay. Previously it was outside of the catch
and this had the effect of removing the overlay even in the case where
the submitForm was successful which meant there was a flash of the
checkout with no overlap before the redirect to the thank you page.

There's no need to remove the overlay unless we're dropping the user
back on the checkout with errors in case something went wrong.