Avoid race condition when updating error state object (#7147)
When autofilling a credit card we observed that occassionally on Chrome
an error would show next to one of the card fields preventing the form
from submitting. I think that was because the call to update the empty
field state was overwriting a previous call from another field. Using
the callback version of the set state prevents this.