Refactor landing page -> checkout e2e smoke tests (#7148)
This is purely a refactoring to DRY up the new e2e tests which begin on the
relevant landing page and navigate through to the checkout and complete a
purchase. So far this has been done for the following smoke tests:

* Ad Lite
* Newspaper
* 3 tier products

I noticed while working on these tests that the one piece which differed
between them was transitioning from the landing page to the checkout. The rest
of the work is common. The actual checkout form work continues to be done by
completeGenericCheckout.

I've added a new function visitLandingPageAndCompleteCheckout which does the
common work, taking a callback to handle the navigation from landing page to
checkout.