Update form-data NPM package from 4.0.2 to 4.0.4
This fixes a dependabot vulnerability.

There were two places where 4.0.2 existed in the lockfile:

In support-frontend:

devDependencies:
jest-environment-jsdom 29.7.0
└─┬ jsdom 20.0.3
  └── form-data 4.0.2
jest-environment-jsdom-global 4.0.0
└─┬ jest-environment-jsdom 29.7.0 peer
  └─┬ jsdom 20.0.3
    └── form-data 4.0.2

In bigquery-acquisitions-publisher:
devDependencies:
vitest 3.1.4
└─┬ jsdom 20.0.3 peer
  └── form-data 4.0.2

In both cases it was being brought in as a transative dependency of
jsdom, which species a version range of ^4.0.0. So the non-vulnerable
version 4.0.4 fits this range. Unfortunately using pnpm update form-data
resulted in lots of unrelated changes so I resorted to updating the
lockfile manually to resolve jsdom's form-data version to 4.0.4 and
deleting the 4.0.2 resolution. Running pnpm install then reports this as
broken and fixes, picking up the desired 4.0.4.