Fix jackson-core vulnerability in supporter-product-data (#7160)
This PR is part 2 of fixing a vulnerablility reported by dependabot for
jackson-core. The vulnerability affects projects which use jackson-core <
2.15.0.

Most of our projects use a later version but not stripe-patrons-data or
supporter-product-data. The vulnerable version of jackson-core is pulled in via
the AWS SDK v1 libraries, so I think the strategy should be to migrate to v2
(v1 is old).

The fix for stripe-patrons data was done in #7113.

This PR migrates the v1 SDK usage in supporter-product-data to v2. The
following v1 SDKs are used by supporter-product-data:

* SSM
* S3
* SQS