Fix /p/<promo code> endpoint (#7187)
Before this was erroring when the promo code specified in the URL path
was for a newspaper product. The reverse route lookup was failing, I
think because uk/subscribe/paper is not a geo aware path, it only exists
for uk.

I've tested locally and now visiting /p/<promo code> redirects to
/uk/subscribe/paper?promoCode=<promo code>, which I think is the
expected behaviour.