Fix unintended submission of PayPal checkout without BAID
In #7173 a change was made to programatically submit the checkout form
in the validate callback in order to get the form validation to kick in
and scroll the user to the relevant form field errors. It looks like
this had an unintended side effect of actually submitting to the backend
in the case that the fields were all valid. However since the user
hasn't yet auth'd with PayPal there's no BAID so the backend correctly
rejects the request. This PR changes things to only scroll to the errors
on attempting to submit the form, and only if we know that there are
errors.