Prevent invalid email confirmation validation error when using autofill
When using auto-fill we sometimes see the validity check for the email
confirmation field being equal to the email field fail. I think this is
happening because the state hasn't updated by the time we trigger the
check on blur. Delaying this slightly seems to fix the issue. It isn't
super elegant but I'm not sure of a better way to handle this.