Prevent invalid email confirmation validation error when using autofill (#7197)
When using browser auto-fill we sometimes see the validity check for the email
confirmation field being equal to the email field fail (seems to be Firefox
specific). I think this is happening because the state hasn't updated by the
time we trigger the check on blur. Delaying this slightly seems to fix the
issue. It isn't super elegant but I'm not sure of a better way to handle this.

I chose a delay of 50ms as to my eyes there's no perceptible delay, while still
seems to have the desired affect. Initially I went with 100ms but that was
noticeably slower than no delay.