Fix the first delivery date shown on the thank you page for National Delivery subscriptions (#7216)
The first delivery date shown on the thank you page is incorrect for National
Delivery subs. This was noticed by Kent. It turns out the function we use to
calculate the date was using old voucher logic for National Delivery. We didn't
handle National Delivery explicitly so this was falling through to the default
(which was voucher logic).

I've tidied up some of the tests around this date logic, and added a new one
for the bug this fixes.

There's a lot more tidy up and consolidation to be done. There's a lot of
fragmented date calculation logic in the codebase which could be consolidated.
Voucher logic can be removed since we no longer sell vouchers through the
support site.