Remove Aus/UTS specific elements from global student landing page (#7224)
The current implementation of the global student landing page re-uses
components from the UTS/Aus student landing page, which has copy and the image
hardcoded. This PR changes the components to be flexible and usable in both
contexts (global and UTS student offer) by moving away from hard coded copy in
the building block components (like StudentHeader) to passing this data in as
props. The following elements of the landing page are now configured
differently on the Aus landing page vs the global one:

* Header image
* Header text
* Sub header text