We don't want to trigger the create recurring product fail alarm for test users in prod
When the current user is a test user, emit the metric with a stage of
CODE rather than PROD. This will not trigger the alarm which is the
behaviour we want. In prod test users use CODE APIs. We don't want a
CODE API call failing to trigger this alarm.