Adapt SupporterPlus email fields if the rate plan is fixed term (#7249)
This is because in the email which these fields are used by we want to
modify the copy depending on whether the product will auto renew.

This currently affects two email fields:

* subscription_details (existing field)
* is_fixed_term (new field)

Currently the only fixed term S+ rate plan is OneYearStudent.

When this lambda is migrated to TypeScript we should use the product
catalog (which has term type) instead of using a hard coded list of
fixed term rate plans. We're not doing that now because we think the
effort required to work with the product catalog in Scala is too great.