Fix thank you page title for Guardian paper products (#7267)
The old logic took the rate plan and added a single exception where the
Everyday rate plan was converted to Every day. All other rate plans were
left as-is. The exception was no longer used since we're selling the
Plus versions of the rate plans now. So you'd see copy like:

"You have now subscribed to the EverydayPlus package"

It was pointed out that we shouldn't be showing the Plus bit, and when
the rate plan name is two words they should be displayed as such. E.g.
Everyday -> Every day, Sixday -> Six day.

This mapping already exists in productCatalogDescription in
productCatalog.ts and I think we should be using this. I've removed the
work "package" from the template as it already appears in the product
catalog data.

Note that the Observer uses different header copy on the thank you page,
and this doesn't affect that.