Don't trigger the create recurring product fail alarm for test users in prod (#7247)
When the current user is a test user, emit the ServerSideCreateFailure metric
with a stage of CODE rather than PROD.

This change means the prod alarm won't be triggered by test users which is the
behaviour we want. In prod test users use CODE APIs. We don't want a CODE API
call failing to trigger this alarm.

I've verified this change by breaking one of the server side API calls
(delivery coverage) to trigger the ServerSideCreateFailure to be emitted. In
CODE a stage of CODE is used for both test users and non test users, as
expected. In Dev when you're a test user a stage of CODE is used, when not a
test user a stage of DEV is used. This shows that the distinction between test
users and non test users is being made correctly. (Yes, it's slightly odd for
the stage to be CODE when acting as a test user in DEV, but it's quite an
unlikely scenario and it's useful to be able to see the distinction).