Only install Chromium for playwright tests (#7274)
In the cron and smoke e2e test GHA setup, only install the chromium Playwright
browser, because it's the only one we use.

It makes the setup more efficient and faster. The timings seem to vary a bit
throughout the day, but this seems to have a good speed benefit and faster
build checks is always nicer!