Update image on collection tab on new paper landing page (#7288)
Before both the collection and home delivery tabs were using the same image
(the collection one). This PR updates the code to show a different image on
each tab.

Note: there's a bit of inconsistency between the home delivery image and the
collection image. The latter has a transparent background but the former
doesn't.

I renamed things a little bit - I don't think the old name of
guardianWeeklyTabHero made sense.