Only add tab anchor to URL when the user selects a tab (#7295)
On the new print+ landing page, only add the tab anchor parameter to the URL
when the user actively selects a tab, not by default when the page loads.

Previously when windowSetHashProperty was called from the useEffect this ran
whenver the tab was changed but also on the initial render. This had the effect
of adding and scrolling to the tab section whenever the user lands on the page.
Instead, do this when a user actively selects a tab.