Enable OK actions for product acquisition alarms (#7317)
These are the alarms which are of the format "No successful x in y
hours" which we have for most product/payment method combinations.
Usually if one goes off the first thing to check is whether it's back to
OK already (for example if we see one of these overnight or over the
weekend). This will surface that information in the alarms channel.