Update some generic checkout form field limits (#7318)
Reduce max length of first name to 30 characters. First names longer than 30
are rejected by GoCardless (via Zuora). We could make this conditional based on
payment method since there's a larger limit when not using GoCardless but I
think the additional complexity isn't worth it.

Add limit to phone field to reflect limit in backend systems. There's a limit
in Zuora and Salesforce, so let's reflect that here (though I think it's
unlikely we'd ever hit this with a legit phone number).

Also contains a refactoring to be more explicit about when to show phone field.
We only show it for Guardian Weekly gift, and we already have a boolean prop to
signal this, so use the prop.