Fix Guardian Weekly checkout promo terms link (#7309)
Update the promo terms link on the Guardian Weekly checkout to reflect the
promo being used. I've also updated the promo terms page itself to remove some
old hardcoded pricing/saving information.

This was initially flagged because the link was using the wrong domain. This
caused us to notice that the (hardcoded) promo code in the link was out of
date. Instead, dynamically build the promo terms link to reflect the currently
applied promo code (if any).

Also, link using just the path so it goes to the right place regardless of
stage.