Migrate storybook to use vite as the build tool (#7353)
This PR updates Storybook to use Vite as the build tool instead of Webpack.

Storybook 9 removes support for building Preact components with Webpack so we
have to switch to Vite before we do that upgrade.

The aliases required in the vite config are more verbose than in the webpack
config, I couldn't find a way to make these work more generically.

I've removed the special .pasteupimport bit of the pipeline by inlining the
palette colours (we'd like to move away from this to source colours anyway,
this is using an old approach).