Observer branded holding content (#7466)
This PR fixes the flash of blue Guardian branded background on Observer branded
pages. Before this PR there was a single component HoldingContent which
rendered a skeleton Guardian branded page. This was used in two places:

* server side via the build-ssr script, which compiles the tsx to js and is
  used server side in the Play Twirl template
* client side as a fallback in the router via Suspense

This PR introduces a new component ObserverHoldingContent branded with the
Observer bg colour and logo, which we use on the observer subdomain. To reduce
duplication I've used the ObserverLayout component inside the
ObserverHoldingContent component. This required some changes to code which
assumed it would only ever render client side (mostly access to the window
object).

Load any Observer branded page (for example the checkout in dev). You should
not see a flash of blue. To verify the server rendered content is correct you
can load the page without JS enabled.

This change does increase the router bundle size since both layouts are
currently included, though there is a corresponding reduction in the bundle
size of the thank you and checkout pages (which themselves previously contained
both layouts). I experimented with a few things to reduce this, without much
success. My plan is to push this out and keep an eye on Core Web Vitals
reporting to measure whether there is an impact on users.