Fix Observer checkout for signed in users
Before this change, when the user is signed in the oauth redirect flow
to profile.theguardian.com is currently broken on the
observer.theguardian.com subdomain. This is because the redirect_uri is
not correctly set (currently from the observer subdomain it's set to the
support subdomain).

This PR fixes things to follow the example of live.theguardian.com and
use a different value from config to know where to redirect back to.