Don't show discount vs retail when there's a promo code (#7516)
On the newspaper subscribe landing page, if a newspaper subscription package
has a promo code, don't show the % saving vs retail for the promo code.

We're concerned this could be confusing for customers, as we don't make it
clear yet how long the promo code discount will apply for. In future we'll add
this to the card.