Upgrade node-forge to 1.3.2
Dependabot reported a vulnerability.

I wanted to run:

$ pnpm --filter support-frontend up node-forge

But this unexpectedly touched lots of other dependencies. So I instead
ran:

$ pnpm --filter support-frontend add node-forge@1.3.2

This gets us to a non-vulnerable version but also added it to the
support-frontend package.json (which we don't want - it's not a direct
dependency), so I removed it from there leaving only the changes in the
lockfile.